+++ crates/shared/src/arguments.rs
@@ -3,11 +3,9 @@
 
 use {
     crate::{
-        account_balances,
         gas_price_estimation::GasEstimatorType,
         price_estimation::PriceEstimators,
         rate_limiter::RateLimitingStrategy,
-        signature_validator,
         sources::{
             balancer_v2::BalancerFactoryKind,
             uniswap_v2::UniV2BaselineSourceParameters,
@@ -146,12 +144,6 @@ pub struct Arguments {
 
     #[clap(flatten)]
     pub tenderly: tenderly_api::Arguments,
-
-    #[clap(flatten)]
-    pub balances: account_balances::Arguments,
-
-    #[clap(flatten)]
-    pub signatures: signature_validator::Arguments,
 
     #[clap(flatten)]
     pub logging: LoggingArguments,
@@ -406,8 +398,6 @@ impl Display for Arguments {
         write!(f, "{}", self.ethrpc)?;
         write!(f, "{}", self.current_block)?;
         write!(f, "{}", self.tenderly)?;
-        write!(f, "{}", self.balances)?;
-        write!(f, "{}", self.signatures)?;
         writeln!(f, "log_filter: {}", self.logging.log_filter)?;
         writeln!(
             f,
